requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
google-auth-oauthlib==1.2.2
google-api-python-client==2.177.0
requests==2.32.4
Streamlit==1.47.1
SQLAlchemy==2.0.42
Werkzeug==3.1.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==2.1.5
click==8.2.1
colorama==0.4.6
blinker==1.9.0
certifi==2024.7.4
charset-normalizer==3.3.2
idna==3.7
urllib3==2.2.1
pyarrow==17.0.0
numpy==1.26.4
pandas==2.2.2
protobuf==4.25.3
attrs==23.2.0
cachetools==5.3.3
flatbuffers==24.3.25b0
google-api-core==2.19.0
google-auth==2.31.0
google-cloud-core==2.4.1
google-crc32c==1.5.0
google-resumable-media==2.1.2
googleapis-common-protos==1.63.0
grpcio==1.64.1
grpcio-status==1.62.1
pyasn1==0.6.0
pyasn1-modules==0.4.0
rsa==4.9
six==1.16.0
tzdata==2024.1
python-dotenv==1.1.1
oauthlib==3.3.1
requests-oauthlib==2.0.0
